Cast string to float to ensure discount line is formatted correctly when greater than zero but less than one.

+++ app/code/Ometria/Api/Controller/V1/Orders.php
@@ -392,7 +392,7 @@ class Orders extends Base
      */
     private function formatLineItemDiscount(string $amount)
     {
-        return (int)$amount > 0 ? sprintf('-%s', $amount) : $amount;
+        return (float)$amount > 0 ? sprintf('-%s', $amount) : $amount;
     }
 
     /**
